+++ AutomatedTellerMachine/src/main/java/ru/sber/projectATM/AutomatedTellerMachine/Main.java
@@ -1,0 +1,12 @@
+package ru.sber.projectATM.AutomatedTellerMachine;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+
+        //SpringApplication.run(Main.class, args);
+    }
+}
