+++ AutomatedTellerMachine/src/main/java/ru/sber/projectATM/AutomatedTellerMachine/User.java
@@ -1,0 +1,23 @@
+package ru.sber.projectATM.AutomatedTellerMachine;
+
+import lombok.Setter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+import ru.sber.projectATM.AutomatedTellerMachine.controller.controllerATM;
+import ru.sber.projectATM.common.*;
+
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.Pattern;
+
+@Setter
+//@AllArgsConstructor
+@Component("user")
+public class User {
+
+
+//    public BalanceResponse getBalance(controllerATM controllerAtm) {
+//        return controllerAtm.getBalance(pan, pin, 1);
+//    }
+}
