+++ bank/src/main/java/ru/sber/projectATM/bank/Main.java
@@ -1,0 +1,11 @@
+package ru.sber.projectATM.bank;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+
+        //SpringApplication.run(Main.class, args);
+    }
+}
