+++ common/src/main/java/ru/sber/projectATM/common/BalanceResponse.java
@@ -1,0 +1,12 @@
+package ru.sber.projectATM.common;
+
+import lombok.Setter;
+import lombok.ToString;
+import ru.sber.projectATM.common.handbook.Status;
+
+@ToString
+@Setter
+public class BalanceResponse {
+    private Balance balance;
+    private Status RC;
+}
