+++ common/src/main/java/ru/sber/projectATM/common/DefaultAccount.java
@@ -1,0 +1,18 @@
+package ru.sber.projectATM.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class DefaultAccount extends Account {
+    private int id;
+    private String PAN;
+    private Balance balance;
+
+    public Balance getDefaultBalance() {
+        return balance;
+    }
+}
