+++ common/src/main/java/ru/sber/projectATM/common/Request.java
@@ -1,0 +1,11 @@
+package ru.sber.projectATM.common;
+
+
+import lombok.Value;
+
+@Value
+public class Request {
+    private final int id;
+    private final String PAN;
+    private final int PIN;
+}
