+++ common/src/main/java/ru/sber/projectATM/common/handbook/Currency.java
@@ -1,0 +1,7 @@
+package ru.sber.projectATM.common.handbook;
+
+public enum Currency {
+    RUR,
+    USD,
+    EUR
+}
