+++ common/src/main/java/ru/sber/projectATM/common/handbook/Status.java
@@ -1,0 +1,7 @@
+package ru.sber.projectATM.common.handbook;
+
+public enum Status {
+    SUCCESS,
+    FAILED,
+    WAITED
+}
