--- src/main/java/ru/sber/projectATM/core/bank/accounting/BalanceResponse.java
@@ -1,12 +1,0 @@
-package ru.sber.projectATM.core.bank.accounting;
-
-import lombok.Setter;
-import lombok.ToString;
-import ru.sber.projectATM.core.handbook.Status;
-
-@ToString
-@Setter
-public class BalanceResponse {
-    private Balance balance;
-    private Status RC;
-}
