--- src/main/java/ru/sber/projectATM/core/bank/accounting/Client.java
@@ -1,18 +1,0 @@
-package ru.sber.projectATM.core.bank.accounting;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.Setter;
-
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class Client<T extends Account> {
-    @NonNull
-    private int id;
-    private T account;
-
-
-}
