--- src/main/java/ru/sber/projectATM/core/handbook/Currency.java
@@ -1,7 +1,0 @@
-package ru.sber.projectATM.core.handbook;
-
-public enum Currency {
-    RUR,
-    USD,
-    EUR
-}
