--- src/main/java/ru/sber/projectATM/core/handbook/Status.java
@@ -1,7 +1,0 @@
-package ru.sber.projectATM.core.handbook;
-
-public enum Status {
-    SUCCESS,
-    FAILED,
-    WAITED
-}
